+++ src/main/java/recsys/evaluator/TopKRecommenderEvaluator.java
@@ -110,7 +110,7 @@ public abstract class TopKRecommenderEvaluator implements RecommenderEvaluator {
 
 		LongPrimitiveIterator it = dataModel.getUserIDs();
 		while (it.hasNext()) {
-			long userID = it.nextLong();
+			long userID = it.nextInt();
 			if (random.nextDouble() < evaluationPercentage) {
 				splitOneUsersPrefs(trainingPercentage, trainingPrefs, testPrefs, userID, dataModel);
 			}
